move old cmds
