+++ MEClass3/old_commands/plot_entry.py
@@ -1,5 +1,5 @@
 
-import MEClass3.plot_performance
+import MEClass3.old_commands.plot_performance
 
 def main():
     MEClass3.plot_performance.main()
